API Reference

MR data types

MagneticResonanceSignals.MRExperiment — Type

load_twix(twix_file_name)
load_twix(io)

MRExperiment(metadata, coils, quality_control, acq_data)

Container for data from a generic magnetic resonance experiment: a series of Acquisitions, each of which records the coil response due to induced nuclear magnetization. This is currently specialized for the type of raw data produced by a Siemens scanner but could be generalized.

Note that the input pulse sequence is not recorded in Siemens raw format, so it's not available here. The only way to know this in full detail is to simulate the sequence with the same input parameters using the Siemens poet tool from the Siemens proprietary IDEA development environment.

MagneticResonanceSignals.PRESS — Type

A standard Point Resolved Spectroscopy (PRESS) experiment with num_averages acquisitions, possibly with navigator and reference scans included.

This is implemented as "single voxel spectroscopy" and is the standard product sequence for in vivo MR spectroscopy on Siemens scanners as of early 2020.

MagneticResonanceSignals.LCOSY — Type

An L-COSY experiment with (num_averages × nsamp_t1) acquisitions, possibly with reference scans included.

The L-COSY pulse sequence is described in:

Thomas, M. Albert, et al. "Localized two‐dimensional shift correlated MR spectroscopy of human brain." Magnetic Resonance in Medicine 46.1 (2001): 58-67.

Thomas, M. Albert, et al. "Evaluation of two‐dimensional L‐COSY and JPRESS using a 3 T MRI scanner: from phantoms to human brain in vivo." NMR in Biomedicine 16.5 (2003): 245-251.

Querying MR data

MagneticResonanceSignals.meta_search — Function

meta_search(metadata, pattern)

Search through MR experiment metadata for a given regular expression, pattern or for a case insensitive string pattern.

MagneticResonanceSignals.standard_metadata — Function

Extract standard MR metadata about an experiment

MagneticResonanceSignals.scanner_time — Function

scanner_time(acq | expt)

Get internal scanner time stamp from an acquisition or sequence of timestamps from an experiment.

MagneticResonanceSignals.sampledata — Function

sampledata(expt, index; downsample=1)

Return the acquired data from expt at a given acqusition index. If downsample>1, the data will be subsampled by the given rate by truncating the tails of the signal in the Fourier spectral domain. This has the effect of removing noise by filtering away irrelevant high and low frequency components.

MagneticResonanceSignals.count_cycles — Function

count_cycles(twix::MRExperiment)

Get phase cycle length for Siemens twix.

High level file IO and conversion

MagneticResonanceSignals.mr_load — Function

mr_load(data)

High level function for loading MR data and recognizing which experiment was run. The returned object describes the experimental conditions and acquisition schedule which can be used during data processing. data may be a file name, IO stream, or a data structure describing raw acquisition data, for example, Siemens Twix.

If the experiment is not recognized, throw an exception.

mr_load(twix::MRExperiment, repair=false)

If repair=true, attempt to repair the experiment by discarding a subset of broken acquisitions where possible.

Recognizes spectro experiments in Siemens twix format, producing one of:

PRESS (Siemens product sequence SVS_SE)

L-COSY (Customer sequences srcosy, svs_lcosy)

MagneticResonanceSignals.twix_to_nmrpipe — Function

twix_to_nmrpipe(twix, nmrpipe_path)

twix is a path or open IO stream to the original twix data file. nmrpipe is a path or open IO stream to the result file to be written.

Low level file IO

MagneticResonanceSignals.load_rda — Function

load_rda(rda_file)

Load header dictionary and spectroscopy data from a Siemens .rda file. Returns header,data, where header is a simple dictionary containing the ASCII header, and data is the time domain induced magnetization signal.

Single voxel spectroscopy data will be loaded as a vector containing the single magnentization signal.

For spectroscopy with a spatial component (that is, CSI data with at least one header field CSIMatrixSize[i] not equal to 1), an array of size N×I×K×J will be returned where I,J,K are the CSI indices 0,1,2 and the FID length is N.

MagneticResonanceSignals.save_rda — Function

save_rda(file, header, data)

Save a single FID as Siemens format .rda. header should be a dictionary of key values for the rda ASCII header.

MagneticResonanceSignals.load_twix — Function

load_twix(filename; header_only=false, acquisition_filter=(acq)->true,
          meas_selector=last)

Load raw Siemens twix ".dat" format, producing an MRExperiment containing a sequence of acqisitions.

For large files, acquisitions may be filtered out during file loading by providing a predicate acquisition_filter(acq) which returns true when acq should be kept. By default, all acquisitions are retained.

N4 VD-version twix may have data from more than one sequence (for example, the tune-up data may be included). Normally the data you're looking for is in the meas_selector=last chunk, but meas_selector is provided for the cases where you want to select other measurements.

MagneticResonanceSignals.save_felix — Function

save_felix(fname, data; bandwidth, frequency)

Create Felix NMR file for 2D NMR experiment with 2D data matrix size npoints = size(data), spectral width bandwidth = (t2_bw, t1_bw) and given spectrometer frequency, which should be unit-compatible with Hz.

data[i,:] is assumed to contain the real time FIDs as acquired in the standard 2D COSY experiment (ie, the "t2" dimension).

MagneticResonanceSignals.save_nmrpipe — Function

save_nmrpipe(io, signal, axes; frequency, ref_freq_offset)

Convert twix into NMR Pipe format to be loaded into third party analytical software, such as MNova. This currently supports up to 2 dimension.

At minimal, for each dimension we should have:

Signal data

Axes

Observation frequency (frequency)

Carrier frequency (ref_freq_offset)

ref_freq_offset is a tuple containing the relative frequency offset between obs_freq and frequency which will be set to 0 on the ppm scale.

High level signal processing

MagneticResonanceSignals.extract_fids — Function

extract_fids(expt; downsample=1)

Extract raw fid from experiment (PRESS/LCOSY) and apply channel combination

MagneticResonanceSignals.simple_averaging — Function

simple_averaging(spectro_expt)

Simple channel combination and averaging for spectroscopic data acquisition.

Note that this uses a simple mean to combine acquisitions across the phase cycling dimension; it does no frequency alignment or other calibration.

simple_averaging(fids)

Do simple averaging from FIDs

MagneticResonanceSignals.spectrum — Function

spectrum(lcosy::LCOSY,
         win1=t->sinebell(t, pow=2),
         win2=t->sinebell(t, skew=0.3, pow=2),
         t1pad=4,
         downsample=1)

Compute spectrum from lcosy data with "standard" L-COSY processing parameters as have traditionally been used by Mountford et. al.

Simple averaging

T2: skewed sine bell squared window, skew=0.3

T1: Sine bell squared window, 4x zero padded

spectrum(press::PRESS,
         win=t->sinebell(t, pow=2),,
         t1pad=4,
         downsample=1)

Compute spectrum from press data with processing parameters as follows.

Simple averaging

T: Sine bell squared window, 4x zero padded

spectrum(signal::AxisArray)

Compute the spectrum from time domain signal via Fourier Transform.

Low level signal processing

Channel combination

MagneticResonanceSignals.pca_channel_combiner — Function

pca_channel_combiner(signals; signal_range=nothing, channels=:)

Compute channel combination object using PCA, assuming that each acquisition in signals has the same relative geometry of sample and coils. signals[i] is assumed to be an array of shape num_samp × num_chan. We use the first few samples of all acquisition data together in the same calculation to get a good estimate of the correct channel weights.

These assumptions should be valid to assess the relative coil SNR for spectroscopy of a single voxel; for other experiments you may want to use a different method, or restrict calculation of weights to a different signal_range.

channels can be set to a index-like object in order to combine only a subset of channels.

MagneticResonanceSignals.combine_channels — Function

combine_channels(combiner::ChannelCombiner, data)

Combine channels data residing in a NxC matrix data, where N is number of temporal samples and C the number of channels. Alternatively, data can be an acquisition.

See pca_channel_combiner() to create a ChannelCombiner functor.

Windows and windowing

MagneticResonanceSignals.zeropad — Function

zeropad(fid, axis, pad)

Zero pad an FID for Fourier interpolation

MagneticResonanceSignals.apply_window — Function

apply_window(signal, axis1=>win1, axis2=>win2, ...)

Apply window functions win1 along axis1, win2 along axis2, etc. The windows should be functions over the dimensionless time domain 0..1; the axes should be of type AxisArrays.Axis.

Background

For the purposes of spectroscopy, the MR signal is generally an oscillating decaying signal in the time domain, with different metabolites having different decay constants. When computing a spectrum with a simple Fourier transform, one may window the time domain data to emphasize one metabolite or another in the resulting spectrum. (Spatially resolved spectroscopy involves some kind of MR echo so the initial part of the signal may be increasing, but the this doesn't affect the irreversible decay mechanisms.)

For qualitative exploration of the spectra, one may therefore want to choose out of several windowing functions with convenient parametric forms. Note that for quantitative measurement of metabolite concentration, windowing distorts the relative peak volumes and it's probably better to choose a time domain fitting method.

Note that these MR-specific reasons are subtly different - and apply in addition to - the generic signal processing reasons for time domain windowing.

MagneticResonanceSignals.apply_window! — Function

Apply windowfunc over the dimensionless time range (0:tlen-1)/tlen) of fid along dimension axis which can be an Axis or Integer.

See also apply_window.

MagneticResonanceSignals.sinebell — Function

sinebell(t; skew=1, pow=1)

The sinebell window on the dimensionless time range t ∈ 0..1.

Use pow=2 for sinebell squared. Setting the skew parameter to something other than the default of 1.0 gives a skewed sinebell window, as in Felix NMR.

Water suppression

MagneticResonanceSignals.hsvd — Function

hsvd(fid, rank)

Use the Hankel singular value decomposition method for fitting a sum of exponential signals of the form

xₙ = Σₖ Aₖ zₖⁿ

to compute the ampltudes Aₖ and complex-valued poles zₖ. Any poles with |zₖ| > 1 are discarded to avoid numerical instability during the fit. A tuple of poles, amplitudes and the exponential basis (z,A,basis) is returned.

The discrete-time and continuous time signals are related by

sₙ =  Aₖ zₖⁿ  =  Aₖ exp((-bₖ + iωₖ) * tₙ)

or

zₖ  =  exp((-bₖ + iωₖ) * Δt)

where ωₖ is an angular frequency and bₖ the decay rate in Hz.

Ref. to NMR literature: Barkhuijsen et al., "Improved Algorithm for Noniterative Time-Domain Model Fitting to Exponentially Damped Magnetic Resonance Signals", J. Mag. Res., (1987) https://doi.org/10.1016/0022-2364(87)90023-0

More broadly these techniques seem to come under the title "Singular Spectrum analysis".

MagneticResonanceSignals.hsvd_water_suppression — Function

function hsvd_water_suppression(fid; dt = 0.0008, water_center_freq=0.0,
                                bandwidth=100.0, hsvd_rank=50)

Use HSVD to suppress a spectral region.

Phase correction

MagneticResonanceSignals.adjust_phase — Function

function adjust_phase(spectrum, zero_phase=0.0, first_phase=0.0, fixed_frequency=0.0)

Adjust phases of the signal.

Example

expt = mr_load("path/to/twix")
spec = spectrum(expt)
ph0, ph1 = ernst(spec)
spec_ph = adjust_phase(spec; zero_phase=ph0, first_phase=ph1)

MagneticResonanceSignals.ernst — Function

function ernst(spectrum)

Estimates the zero and first order phase parameters which minimise the integral of imaginary part of the spectrum.

This implementation is based on suspect.py. See: https://github.com/openmrslab/suspect/blob/master/suspect/processing/phase.py

Baseline correction

MagneticResonanceSignals.baseline_als — Function

function baseline_als(spectrum, lambda::Float64, p::Float64; niter::Int64=10)

Baseline correction based on ALS (Asymmetric Least Square).

Lambda is 2nd derivative constraint, which contributes to the smoothness, while p is weighting of positive residuals, which affect to the asymmetry.

This implementation is inspired by https://stackoverflow.com/a/29185844/5023889

Plotting

MagneticResonanceSignals.felix_colors — Constant

A popular color map for multidimensional NMR data as defined in Felix NMR